% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxmod.R
\name{maxmod}
\alias{maxmod}
\title{minmod
fonction retournant la modalité avec le plus d'occurence.}
\usage{
maxmod(data, variable)
}
\arguments{
\item{data}{un dataframe}

\item{variable}{une variable qualitative de ce dataframe}
}
\description{
minmod
fonction retournant la modalité avec le plus d'occurence.
}
