% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmod.R
\name{minmod}
\alias{minmod}
\title{minmod
fonction retournant la modalité avec le moins d'occurence.}
\usage{
minmod(data, variable)
}
\arguments{
\item{data}{un dataframe}

\item{variable}{une variable qualitative de ce dataframe}
}
\description{
minmod
fonction retournant la modalité avec le moins d'occurence.
}
