% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scinde.R
\name{scinde}
\alias{scinde}
\title{scinde une data
fonction permettant de recuperer une subdata avec seulement le type de variables specifie}
\usage{
scinde(data, type = NULL)
}
\arguments{
\item{data}{un dataframe}

\item{type}{une chaine de caractere specifiant la partie retourne.
Si null,renvoie une liste avec les noms des variables pour chaque type.}
}
\description{
scinde une data
fonction permettant de recuperer une subdata avec seulement le type de variables specifie
}
